refactoring -> storing float instead of int

+++ src/main/java/org/sankozi/rogueland/model/Actor.java
@@ -15,7 +15,7 @@ public interface Actor extends Destroyable{
     Damage getPower();
 
     /** Returns value stored in int */
-    public int actorParam(Param param);
+    public float actorParam(Param param);
 
     public enum Param {
         DAMAGE_MIN,
